refactoring: nuget command refactored
